travistooling: rename git.py to git_utils.py for disambiguation
