+++ travistooling/__init__.py
@@ -1,8 +1,3 @@
 # -*- encoding: utf-8
 
-import subprocess
-
-
-# Root of the Git repository
-ROOT = subprocess.check_output([
-    'git', 'rev-parse', '--show-toplevel']).decode('ascii').strip()
+from travistooling.git_utils import get_changed_paths, git, ROOT
